requirements.txt
# Runtime — manter alinhado ao projeto; instalar com: pip install -r requirements.txt
langgraph>=0.4.0
langgraph-supervisor>=0.0.5
langchain>=0.3.0
langchain-mistralai>=0.2.0
langchain-core>=0.3.0
langfuse>=2.0.0
python-dotenv>=1.0.0
